created home app and moved index, 404 and 500 urls to iit
