+++ templates/_layouts/default.html
@@ -19,7 +19,7 @@
 </head>
 <body>
 <nav class="navbar navbar-expand-md">
-    <a class="navbar-brand" href="{% url 'talks:index' %}">vtalks.net</a>
+    <a class="navbar-brand" href="{% url 'home:index' %}">vtalks.net</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbar-content" aria-controls="navbar-content" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
     </button>
@@ -51,7 +51,7 @@
                     <a class="nav-link disabled" href="{% url 'corporate:privacy' %}">vtalks.net © 2015-{% now "Y" %}</a>
                 </nav>
                 <nav class="nav">
-                    <a class="nav-link" href="{% url 'talks:index' %}">Home</a>
+                    <a class="nav-link" href="{% url 'home:index' %}">Home</a>
                     <a class="nav-link" href="{% url 'corporate:about' %}">About</a>
                     <a class="nav-link" href="{% url 'corporate:help' %}">Help</a>
                     <a class="nav-link" href="{% url 'corporate:contact' %}">Contact</a>
