+++ web/corporate/urls.py
@@ -5,7 +5,7 @@ from .views import ContactView
 from .views import HelpView
 from .views import TermsView
 from .views import PrivacyView
-from .views import TemplateView
+
 
 app_name = 'corporate'
 urlpatterns = [
@@ -14,7 +14,4 @@ urlpatterns = [
     path('help', HelpView.as_view(), name='help'),
     path('terms', TermsView.as_view(), name='terms'),
     path('privacy', PrivacyView.as_view(), name='privacy'),
-
-    path('404', TemplateView.as_view(template_name="404.html"), name='404'),
-    path('500', TemplateView.as_view(template_name="500.html"), name='500'),
 ]
